update: 添加问题“955.删列造序II”的代码和题解 (#1271) 955: AC.cpp - AC,100.00,81.82% (#1270) + word(en)
一遍过喽
