+++ .commitmsg
@@ -1,10 +1,3 @@
-944: AC.cpp+java+py+go+rust(#1268) + word(en) + 1problem1day
+955: AC.cpp - AC,100.00,81.82% (#1270) + word(en)
 
-cpp - WA
-cpp - AC,27.82%,45.86%
-py(oneline) - CE
-py(oneline) - AC,11.04%,70.13%
-go - AC,100.00%,38.89%
-java - AC,84.34%,6.03%
-rust - CE
-rust - AC,33.33%,83.33%
+一遍过喽
