+++ .github/workflows/1problem1day-autoGenIssueEveryday.yml
@@ -45,4 +45,4 @@ jobs:
             --body "By Github Actions | [$id.$title](https://leetcode.cn/problems/$slug/) | $date" \
             --label 题解 \
             --assignee LetMeFly666 \
-            --repo LetMeFly666/LeetCode-test
+            --repo "${{ github.repository }}"
