+++ .github/workflows/1problem1day-closeAndSetDate.yml
@@ -25,7 +25,7 @@ jobs:
           closed_at="${{ github.event.issue.closed_at }}"
           date_cn=$(date -d "$closed_at +8 hours" +%F)
 
-                    echo "Query project item_id via Issue.projectItems..."
+          echo "Query project item_id via Issue.projectItems..."
           item_id=$(gh api graphql -f query='
             query($issue:ID!) {
               node(id:$issue) {
