+++ Codes/0944-delete-columns-to-make-sorted_20251220.cpp
@@ -2,13 +2,12 @@
  * @Author: LetMeFly
  * @Date: 2025-12-20 22:44:28
  * @LastEditors: LetMeFly.xyz
- * @LastEditTime: 2025-12-20 22:46:33
+ * @LastEditTime: 2025-12-20 22:48:32
  */
 #if defined(_WIN32) || defined(__APPLE__)
 #include "_[1,2]toVector.h"
 #endif
 
-// THIS CAN NOT AC
 class Solution {
 public:
     int minDeletionSize(vector<string>& strs) {
@@ -16,7 +15,7 @@ public:
         int ans = 0;
         for (int j = 0; j < m; j++) {
             for (int i = 1; i < n; i++) {
-                if (strs[i][j] <= strs[i - 1][j]) {
+                if (strs[i][j] < strs[i - 1][j]) {  // 好吧，相同相邻字符是合法的
                     ans++;
                     break;
                 }
