+++ Codes/0944-delete-columns-to-make-sorted_20251220.py
@@ -2,11 +2,10 @@
 Author: LetMeFly
 Date: 2025-12-20 22:44:28
 LastEditors: LetMeFly.xyz
-LastEditTime: 2025-12-20 22:52:49
+LastEditTime: 2025-12-20 22:55:49
 '''
 from typing import List
 
-# THIS CAN NOT ACCESS
 class Solution:
     def minDeletionSize(self, strs: List[str]) -> int:
-        return sum(any(strs[i][j] < strs[i-1][j] for i in range(1, len(strs)) for j in range(len(strs))))
+        return sum(any(strs[i][j] < strs[i-1][j] for i in range(1, len(strs))) for j in range(len(strs[0])))
