+++ Codes/0955-delete-columns-to-make-sorted-ii.go
@@ -1,0 +1,5 @@
+package main
+
+func minDeletionSize(strs []string) int {
+    
+}
