+++ Codes/0955-delete-columns-to-make-sorted-ii.java
@@ -1,0 +1,5 @@
+class Solution {
+    public int minDeletionSize(String[] strs) {
+        
+    }
+}
