+++ Codes/0955-delete-columns-to-make-sorted-ii.py
@@ -1,0 +1,5 @@
+from typing import List
+
+class Solution:
+    def minDeletionSize(self, strs: List[str]) -> int:
+
