+++ README.md
@@ -464,6 +464,7 @@
 |0944.删列造序|简单|<a href="https://leetcode.cn/problems/delete-columns-to-make-sorted/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2025/12/20/LeetCode%200944.%E5%88%A0%E5%88%97%E9%80%A0%E5%BA%8F/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/156117851" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/delete-columns-to-make-sorted/solutions/3861886/letmefly-944shan-lie-zao-xu-zi-fu-chuan-7o3go/" target="_blank">LeetCode题解</a>|
 |0946.验证栈序列|中等|<a href="https://leetcode.cn/problems/validate-stack-sequences/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2022/08/31/LeetCode%200946.%E9%AA%8C%E8%AF%81%E6%A0%88%E5%BA%8F%E5%88%97/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/126616819" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/validate-stack-sequences/solution/letmefly-946yan-zheng-zhan-xu-lie-by-tis-st5n/" target="_blank">LeetCode题解</a>|
 |0952.按公因数计算最大组件大小|困难|<a href="https://leetcode.cn/problems/largest-component-size-by-common-factor/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2022/07/30/LeetCode%200952.%E6%8C%89%E5%85%AC%E5%9B%A0%E6%95%B0%E8%AE%A1%E7%AE%97%E6%9C%80%E5%A4%A7%E7%BB%84%E4%BB%B6%E5%A4%A7%E5%B0%8F/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/126069985" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/largest-component-size-by-common-factor/solution/letmefly-952an-gong-yin-shu-ji-suan-zui-4ljri/" target="_blank">LeetCode题解</a>|
+|0955.删列造序II|中等|<a href="https://leetcode.cn/problems/delete-columns-to-make-sorted-ii/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2025/12/21/LeetCode%200955.%E5%88%A0%E5%88%97%E9%80%A0%E5%BA%8FII/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/156134471" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/delete-columns-to-make-sorted-ii/solutions/3862470/letmefly-955shan-lie-zao-xu-iimo-ni-omn-lbuob/" target="_blank">LeetCode题解</a>|
 |0961.在长度2N的数组中找出重复N次的元素|简单|<a href="https://leetcode.cn/problems/n-repeated-element-in-size-2n-array/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2022/05/21/LeetCode%200961.%E5%9C%A8%E9%95%BF%E5%BA%A62N%E7%9A%84%E6%95%B0%E7%BB%84%E4%B8%AD%E6%89%BE%E5%87%BA%E9%87%8D%E5%A4%8DN%E6%AC%A1%E7%9A%84%E5%85%83%E7%B4%A0/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/124897591" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/find-the-closest-palindrome/solution/letmefly-da-an-hou-xuan-by-letmefly-2-k2vn/" target="_blank">LeetCode题解</a>|
 |0966.元音拼写检查器|中等|<a href="https://leetcode.cn/problems/vowel-spellchecker/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2025/09/14/LeetCode%200966.%E5%85%83%E9%9F%B3%E6%8B%BC%E5%86%99%E6%A3%80%E6%9F%A5%E5%99%A8/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/151690858" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/vowel-spellchecker/solutions/3781983/letmefly-966yuan-yin-pin-xie-jian-cha-qi-16uh/" target="_blank">LeetCode题解</a>|
 |0970.强整数|中等|<a href="https://leetcode.cn/problems/powerful-integers/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2023/05/02/LeetCode%200970.%E5%BC%BA%E6%95%B4%E6%95%B0/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/130461751" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/powerful-integers/solutions/2253101/letmefly-970qiang-zheng-shu-by-tisfy-w22m/" target="_blank">LeetCode题解</a>|
