+++ Solutions/Other-English-LearningNotes-SomeWords.md
@@ -1681,8 +1681,13 @@ categories: [自用]
 |buffalo|v. 威吓，欺骗，迷惑<br/>n. 水牛，(北美)野牛|
 |hoe|n. 锄头<br/>v. 锄地|
 |||
+|surveyor|n. (土地)测量员，（建筑物）鉴定人，检验员|
+|||
 |ale|n. (一杯)(麦芽)啤酒|
 |parish|n. 教区，堂区，教区居民|
+|||
+|imprison|v. 监禁，束缚，关押，入狱|
+|satisfactorily|adv. 满意地，可靠地，顺心|
 
 + 这个web要是能设计得可以闭眼(完全不睁眼)键盘控制背单词就好了。
 + 也许可以加个AI用最近词编故事功能(返回接口中支持标注所使用单词高亮?)
