--- toSay.md
@@ -1,9 +1,0 @@
-<!--
- * @Author: LetMeFly
- * @Date: 2025-12-20 11:14:30
- * @LastEditors: LetMeFly.xyz
- * @LastEditTime: 2025-12-20 11:14:30
--->
-github workflows不能写到子文件夹下！
-
-根本不会递归遍历
